feat(map): the YAML files must not use any Jinja
For each `.yaml` file that `map.jinja` try to load, a `.yaml.jinja`
Jinja template is tried right after to permit jinja manipulation of
values and circumvent yamllint errors when using Jinja in YAML files.

At the end of the load of YAML files and Jinja YAML templates,
`map.jinja` include an optional `post-map.jinja` for post-processing
the `mapdata` variable.